+++ lab1/code/qr.h
@@ -1,17 +1,17 @@
 #pragma once
 
-double* qr(int& n, double*& A, double*& b)
+float* qr(int& n, float*& A, float*& b)
 {
     for (auto i = 0; i < n; ++i)
     {
         for (auto j = i + 1; j < n; ++j)
         {
-            double* T = rotMat(i, j, A, n);
+            float* T = rotMat(i, j, A, n);
             multMat(T, A, n, n, n, n);
             multMat(T, b, n, n, n, 1);
             delete[] T;
         }
     }
-    double* X = gauss(n, A, b);
+    float* X = gauss(n, A, b);
     return X;
 }
